+++ lib/MIP/Cli/Mip/Analyse/Rd_dna.pm
@@ -446,6 +446,17 @@ q{gatk_baserecalibration_known_sites, gatk_haplotypecaller_snp_known_set, gatk_v
     );
 
     option(
+        q{markduplicates_picardtools_opt_dup_dist} => (
+            cmd_aliases   => [qw{ mdpodd }],
+            cmd_flag      => q{picard_mdup_odd},
+            cmd_tags      => [q{Default: 2500}],
+            documentation => q{Picardtools markduplicates optical duplicate distance},
+            is            => q{rw},
+            isa           => Int,
+        )
+    );
+
+    option(
         q{markduplicates_sambamba_markdup} => (
             cmd_aliases   => [qw{ mdsmd }],
             cmd_flag      => q{sambamba_markdup},